# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Simple Web Blog"
  version: "1.0.0"
  title: "simple web blog"
  termsOfService: "https://swagger.io/terms/"
  contact:
    email: "nanzh@mail2.sysu.edu.cn"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/serviece/Test/1.0.0"
...
schemes:
- "https"
paths:
...
  /user/article/{id}/comments:
    get:
      tags:
      - "comment"
      summary: "Get the comments of the article"
      description: "Get the comments of the article"
      operationId: "GetCommentsOfArticle"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID of the article"
        required: true
        type: "integer"
        x-exportParaName: "Id"
        x-exportParamName: "Id"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/Comments"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/response_400"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/response_404"
    post:
      tags:
      - "comments"
      summary: "Create a new comment"
      description: "Create a new comment"
      operationId: "CreateComment"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID of the article"
        required: true
        type: "integer"
        x-exportParaName: "Id"
        x-exportParamName: "Id"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body"
        x-exportParaName: "Body"
        x-exportParamName: "Body"
      responses:
        "200":
          description: "Successful"
          schema:
            $ref: "#/definitions/Comment"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/response_400"
definitions:
...
  Comment:
    type: "object"
    required:
    - "articleId"
    - "author"
    - "content"
    - "date"
    properties:
      date:
        type: "string"
      content:
        type: "string"
      author:
        type: "string"
      articleId:
        type: "integer"
    example:
      date: "date"
      author: "author"
      articleId: 0
      content: "content"
  Comments:
    properties:
      contents:
        type: "array"
        items:
          $ref: "#/definitions/Comment"
    example:
      contents:
      - date: "date"
        author: "author"
        articleId: 0
        content: "content"
      - date: "date"
        author: "author"
        articleId: 0
        content: "content"
  body:
    type: "object"
    properties:
      content:
        type: "string"
      author:
        type: "string"
...
  response_400:
    type: "object"
    properties:
      error:
        type: "string"
        example: ""
  response_404:
    type: "object"
    properties:
      error:
        type: "string"
        example: ""